decals just in case

+++ Assets/!Materials/Projectors/ImageProjector.shader
@@ -19,19 +19,19 @@
  
          // User-specified properties
          uniform sampler2D _ShadowTex; 
-         float4 _Color;
+         fixed4 _Color;
  
          // Projector-specific uniforms
-         uniform float4x4 _Projector; // transformation matrix 
+         uniform fixed4x4 _Projector; // transformation matrix 
             // from object space to projector space 
  
           struct vertexInput {
-            float4 vertex : POSITION;
+            fixed4 vertex : POSITION;
             float3 normal : NORMAL;
          };
          struct vertexOutput {
-            float4 pos : SV_POSITION;
-            float4 posProj : TEXCOORD0;
+            fixed4 pos : SV_POSITION;
+            fixed4 posProj : TEXCOORD0;
                // position in projector space
          };
  
@@ -45,12 +45,12 @@
          }
  
  
-         float4 frag(vertexOutput input) : COLOR
+         fixed4 frag(vertexOutput input) : COLOR
          {
             if (input.posProj.w > 0.0) // in front of projector?
             {
                //return tex2D(_ShadowTex , input.posProj.xy / input.posProj.w); 
-               float4 f = tex2D(_ShadowTex , input.posProj.xy / input.posProj.w);
+               fixed4 f = tex2D(_ShadowTex , input.posProj.xy / input.posProj.w);
                f.rgb *= _Color.rgb;
                return f;
 
@@ -59,7 +59,7 @@
             }
             else // behind projector
             {
-               return float4(0.0, 0.0, 0.0, 0.0);
+               return fixed4(0.0, 0.0, 0.0, 0.0);
             }
          }
